refactor(tag): remove TagService interface
